+++ SportStore/DataWindow.xaml.cs
@@ -3,7 +3,6 @@ using SportStore.Models;
 using System;
 using System.Linq;
 using System.Windows;
-using System.Windows.Controls;
 using System.Windows.Input;
 
 namespace SportStore
@@ -13,8 +12,6 @@ namespace SportStore
     /// </summary>
     public partial class DataWindow : Window
     {
-
-        public static DataGrid datagrid;
         public DataWindow()
         {
             InitializeComponent();
@@ -22,61 +19,67 @@ namespace SportStore
             using (SportStoreContext db = new SportStoreContext())
             {
                 usersGrid.ItemsSource = db.Users.Include(u => u.Role).ToList();
+                Proxy.userGrid = usersGrid;
             }
         }
 
         private void SelectUser(object sender, MouseButtonEventArgs e)
         {
             User user = (User)usersGrid.SelectedItem;
-
             int Id = (usersGrid.SelectedItem as User).Id;
-
-            new UpdateWindow(user, Id).ShowDialog();
+            new AddEditWindow(user, Id).ShowDialog();
+        }
+        private void UpdateUsers_Button(object sender, RoutedEventArgs e)
+        {
+            try
+            {
+                User user = (User)usersGrid.SelectedItem;
+                int Id = (usersGrid.SelectedItem as User).Id;
+                new AddEditWindow(user, Id).ShowDialog();
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
         }
 
         private void AddUsers(object sender, RoutedEventArgs e)
         {
-            new AddWindow(null).ShowDialog();
-
-            using (SportStoreContext db = new SportStoreContext())
-            {
-                usersGrid.ItemsSource = db.Users.Include(u => u.Role).ToList();
-            }
-        }
-
-        private void UpdateUsers(object sender, RoutedEventArgs e)
-        {
-            User user = (User)usersGrid.SelectedItem;
-
-            int Id = (usersGrid.SelectedItem as User).Id;
-
-            new UpdateWindow(user, Id).ShowDialog();
+            new AddEditWindow(null, 0).ShowDialog();
         }
 
         private void Delete(object sender, RoutedEventArgs e)
         {
-            using (SportStoreContext db = new SportStoreContext())
+            var selectedUsers = usersGrid.SelectedItems.Cast<User>().ToList();
+
+            if (MessageBox.Show($"Вы точно хотите удалить {selectedUsers.Count()} пользователей", "Внимание!",
+                 MessageBoxButton.YesNo, MessageBoxImage.Question) == MessageBoxResult.Yes)
             {
-                User user = (User)usersGrid.SelectedItem;
-                if (user is null)
-                    return;
-
-                if (MessageBox.Show($"Вы точно хотитие удалить запись {user.Name}?", "Внимание", MessageBoxButton.YesNo, MessageBoxImage.Question) == MessageBoxResult.Yes)
+                try
                 {
-                    try
-                    {
-                        db.Users.Remove(user);
-                        db.SaveChanges();
-                        MessageBox.Show($"Пользователь {user.Name} удален");
-                    }
-                    catch (Exception ex)
-                    {
-                        MessageBox.Show(ex.Message);
-                    }
+                    SportStoreContext db = new SportStoreContext();
+                    db.Users.RemoveRange(selectedUsers);
+                    db.SaveChanges();
+                    usersGrid.ItemsSource = db.Users.ToList();
+                    MessageBox.Show("Пользователи удалены!");
                 }
-
-                usersGrid.ItemsSource = db.Users.Include(u => u.Role).ToList();
+                catch (Exception ex)
+                {
+                    MessageBox.Show($"{ex.Message}");
+                }
             }
+        }
+
+        private void Window_KeyDown(object sender, KeyEventArgs e)
+        {
+            //    if (e.Key == Key.F5)
+            //    {
+            //        using (SportStoreContext db = new SportStoreContext())
+            //        {
+            //            usersGrid.ItemsSource = db.Users.ToList();
+            //            Proxy.userGrid = usersGrid;
+            //        }
+            //    }
         }
     }
 }
