+++ SportStore/Proxy.cs
@@ -1,0 +1,9 @@
+﻿using System.Windows.Controls;
+
+namespace SportStore
+{
+    public class Proxy
+    {
+        public static DataGrid userGrid { get; set; }
+    }
+}
