--- SportStore/UpdateWindow.xaml
@@ -1,40 +1,0 @@
-﻿<Window x:Class="SportStore.UpdateWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SportStore"
-        mc:Ignorable="d"
-        Title="Update" Height="400" Width="300"
-        Icon="Resources/img1.png"
-        WindowStartupLocation="CenterScreen">
-
-    <Grid>
-        <Border BorderThickness="1"
-                BorderBrush="BlueViolet"
-                Margin="10">
-
-            <StackPanel Margin="5">
-                <TextBlock Text="Имя" Width="200" Height="20" />
-                <TextBox x:Name="textBoxName" Text="{Binding Name}" Width="200" Height="20"/>
-
-                <TextBlock Text="Фамилия" Width="200" Height="20"/>
-                <TextBox x:Name="textBoxSurname" Text="{Binding Surname}" Width="200" Height="20" />
-
-                <TextBlock Text="Отчество" Width="200" Height="20"/>
-                <TextBox x:Name="textBoxPatronymic" Text="{Binding Patronymic}" Width="200" Height="20" />
-
-                <TextBlock Text="Логин" Width="200" Height="20"/>
-                <TextBox x:Name="textBoxLogin" Text="{Binding Login}" Width="200" Height="20" />
-
-                <TextBlock Text="Пароль" Width="200" Height="20"/>
-                <TextBox x:Name="textBoxPassword" Text="{Binding Password}" Width="200" Height="20" />
-
-                <TextBlock Text="Роль" Width="200" Height="20"/>
-                <TextBox x:Name="textBoxRole" Text="{Binding Role.Name}" Width="200" Height="20"/>
-
-                <Button Content="Изменить" Click="UpdateUserButton" Margin="0, 25, 0, 0" Background="#FF76E383" Width="200" Height="20" />
-            </StackPanel>
-        </Border>
-    </Grid>
-</Window>
